% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptasticsearch.R
\name{doc_shared}
\alias{doc_shared}
\title{NULL Object For Common Documentation}
\arguments{
\item{es_host}{A string identifying an Elasticsearch host. This should be of the form 
\code{[transfer_protocol][hostname]:[port]}. For example, \code{'http://myindex.thing.com:9200'}.}

\item{es_index}{The name of an Elasticsearch index to be queried.}
}
\description{
This is a NULL object with documentation so that later functions can call
inheritParams
}
